[Chore] tauri config file 환경별 분리

+++ src-tauri/src/domains/common/model.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct CommandResponse<T> {
     pub success: bool,
-    pub code: i8,
+    pub code: i16,
     pub message: String,
     pub data: Option<T>,
 }
